Move ScriptSender and ReverseInterface to new "control" module
It is not really plain communication-related but rather we implemented
a control protocol here. In my opinion it makes sense to move this into a new
"module", as this will be extended with the changes regarding Cartesian
control and trajectory forwarding.
