+++ CMakeLists.txt
@@ -19,6 +19,8 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -pedantic")
 add_library(urcl SHARED
     src/comm/tcp_socket.cpp
     src/comm/tcp_server.cpp
+    src/control/reverse_interface.cpp
+    src/control/script_sender.cpp
     src/primary/primary_package.cpp
     src/primary/robot_message.cpp
     src/primary/robot_state.cpp
