+++ include/ur_client_library/control/reverse_interface.h
@@ -37,7 +37,7 @@
 
 namespace urcl
 {
-namespace comm
+namespace control
 {
 /*!
  * \brief Control modes as interpreted from the script runnning on the robot.
@@ -65,23 +65,12 @@ public:
    * \param port Port the Server is started on
    * \param handle_program_state Function handle to a callback on program state changes.
    */
-  ReverseInterface(uint32_t port, std::function<void(bool)> handle_program_state)
-    : client_fd_(-1), server_(port), handle_program_state_(handle_program_state), keepalive_count_(1)
-  {
-    handle_program_state_(false);
-    server_.setMessageCallback(
-        std::bind(&ReverseInterface::messageCallback, this, std::placeholders::_1, std::placeholders::_2));
-    server_.setConnectCallback(std::bind(&ReverseInterface::connectionCallback, this, std::placeholders::_1));
-    server_.setDisconnectCallback(std::bind(&ReverseInterface::disconnectionCallback, this, std::placeholders::_1));
-    server_.setMaxClientsAllowed(1);
-    server_.start();
-  }
+  ReverseInterface(uint32_t port, std::function<void(bool)> handle_program_state);
+
   /*!
    * \brief Disconnects possible clients so the reverse interface object can be safely destroyed.
    */
-  ~ReverseInterface()
-  {
-  }
+  ~ReverseInterface() = default;
 
   /*!
    * \brief Writes needed information to the robot to be read by the URCaps program.
@@ -92,40 +81,7 @@ public:
    *
    * \returns True, if the write was performed successfully, false otherwise.
    */
-  bool write(const vector6d_t* positions, const ControlMode control_mode = ControlMode::MODE_IDLE)
-  {
-    if (client_fd_ == -1)
-    {
-      return false;
-    }
-    uint8_t buffer[sizeof(int32_t) * 8];
-    uint8_t* b_pos = buffer;
-
-    // The first element is always the keepalive signal.
-    int32_t val = htobe32(keepalive_count_);
-    b_pos += append(b_pos, val);
-
-    if (positions != nullptr)
-    {
-      for (auto const& pos : *positions)
-      {
-        int32_t val = static_cast<int32_t>(pos * MULT_JOINTSTATE);
-        val = htobe32(val);
-        b_pos += append(b_pos, val);
-      }
-    }
-    else
-    {
-      b_pos += 6 * sizeof(int32_t);
-    }
-
-    val = htobe32(toUnderlying(control_mode));
-    b_pos += append(b_pos, val);
-
-    size_t written;
-
-    return server_.write(client_fd_, buffer, sizeof(buffer), written);
-  }
+  bool write(const vector6d_t* positions, const ControlMode control_mode = ControlMode::MODE_IDLE);
 
   /*!
    * \brief Set the Keepalive count. This will set the number of allowed timeout reads on the robot.
@@ -138,36 +94,14 @@ public:
   }
 
 private:
-  void connectionCallback(const int filedescriptor)
-  {
-    if (client_fd_ < 0)
-    {
-      URCL_LOG_INFO("Robot connected to reverse interface. Ready to receive control commands.");
-      client_fd_ = filedescriptor;
-      handle_program_state_(true);
-    }
-    else
-    {
-      URCL_LOG_ERROR("Connection request to ReverseInterface received while connection already established. Only one "
-                     "connection is allowed at a time. Ignoring this request.");
-    }
-  }
+  void connectionCallback(const int filedescriptor);
 
-  void disconnectionCallback(const int filedescriptor)
-  {
-    URCL_LOG_INFO("Connection to reverse interface dropped.", filedescriptor);
-    client_fd_ = -1;
-    handle_program_state_(false);
-  }
+  void disconnectionCallback(const int filedescriptor);
 
-  void messageCallback(const int filedescriptor, char* buffer)
-  {
-    URCL_LOG_WARN("Messge on ReverseInterface received. The reverse interface currently does not support any message "
-                  "handling. This message will be ignored.");
-  }
+  void messageCallback(const int filedescriptor, char* buffer);
 
   int client_fd_;
-  TCPServer server_;
+  comm::TCPServer server_;
 
   static const int32_t MULT_JOINTSTATE = 1000000;
 
@@ -183,7 +117,7 @@ private:
   uint32_t keepalive_count_;
 };
 
-}  // namespace comm
+}  // namespace control
 }  // namespace urcl
 
 #endif  // UR_CLIENT_LIBRARY_REVERSE_INTERFACE_H_INCLUDED
